HTTP error handling, rainbowprint, & code cleaing
404 and 503 error handling to fetch()
-------------------------------------
To improve consistancy, if the URL is invalid, it should modify self.results to contain a status_code.

If the user's network is not connected to the internet, or if IP Quality Score API is down, fetch should handle it gracefully instead of not at all.

rainbowprint
------------
print() will now be able to print standard output in pretty color gradients.

Code cleaning
-------------
Removed unused import statement.
Modified print() to build the string first before printing, to insure that the string has a gradient as opposed to randomly colored lines (substrings).